+++ CHANGELOG.md
@@ -233,4 +233,5 @@
 * Fixed datatype errors since homebridge 1.3.0
 * Fixed observation time issue in weatherunderground since homebridge 1.3.0
 * Fixed error in history service since homebridge 1.3.0
+* Fixed wrong tempeature in Fahrenheit for dew point and apparent temperature
 * Added donation links for homebridge-ui-x